# app/routes/tags_relationship_ratings.py
# Auto-generated from OpenAPI. Edits may be overwritten. Put business logic in app/services.
from typing import Any, Optional, List, Dict
from uuid import UUID
from datetime import date, datetime
from fastapi import APIRouter, Depends, Path, Query, Body
from app.schemas.models import *  # noqa: F403,F401

router = APIRouter(tags=["tags:relationship_ratings"])


@router.delete(
    "/tags/{id}/relationship_ratings",
    
    status_code=200,
)
async def delete_tags__id__relationship_ratings(TagId: Any = None,) -> None:
    """
    Delete relationship ratings for a tag
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationship_ratings.delete_tags__id__relationship_ratings(...)
    raise NotImplementedError("DELETE /tags/{id}/relationship_ratings not implemented yet")

@router.delete(
    "/tags/{id}/relationship_ratings/{rating_id}",
    
    status_code=200,
)
async def delete_tags__id__relationship_ratings__rating_id_(rating_id: UUID,TagId: Any = None,) -> None:
    """
    Delete relationship rating for a tag
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationship_ratings.delete_tags__id__relationship_ratings__rating_id_(...)
    raise NotImplementedError("DELETE /tags/{id}/relationship_ratings/{rating_id} not implemented yet")

@router.get(
    "/tags/{id}/relationship_ratings",
    
    status_code=200,
)
async def get_tags__id__relationship_ratings(TagId: Any = None,) -> None:
    """
    List relationship ratings for a tag
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationship_ratings.get_tags__id__relationship_ratings(...)
    raise NotImplementedError("GET /tags/{id}/relationship_ratings not implemented yet")

@router.get(
    "/tags/{id}/relationship_ratings/{rating_id}",
    response_model=TagRelationshipRatingInput,
    status_code=200,
)
async def get_tags__id__relationship_ratings__rating_id_(rating_id: UUID,TagId: Any = None,) -> TagRelationshipRatingInput:
    """
    Retrieve relationship rating for a tag
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationship_ratings.get_tags__id__relationship_ratings__rating_id_(...)
    raise NotImplementedError("GET /tags/{id}/relationship_ratings/{rating_id} not implemented yet")

@router.patch(
    "/tags/{id}/relationship_ratings",
    
    status_code=200,
)
async def patch_tags__id__relationship_ratings(TagId: Any = None,body: RelationshipRatingUpdate = Body(...),) -> None:
    """
    Apply or update relationship ratings for a tag
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationship_ratings.patch_tags__id__relationship_ratings(...)
    raise NotImplementedError("PATCH /tags/{id}/relationship_ratings not implemented yet")

@router.patch(
    "/tags/{id}/relationship_ratings/{rating_id}",
    response_model=TagRelationshipRatingInput,
    status_code=200,
)
async def patch_tags__id__relationship_ratings__rating_id_(rating_id: UUID,TagId: Any = None,body: Any = Body(...),) -> TagRelationshipRatingInput:
    """
    Update relationship rating for a tag
    OperationId: 
    Auto-generated stub. Add auth/deps if needed.
    """
    # TODO: Implement business logic in app/services and call it here.
    # Example: return await services.tags_relationship_ratings.patch_tags__id__relationship_ratings__rating_id_(...)
    raise NotImplementedError("PATCH /tags/{id}/relationship_ratings/{rating_id} not implemented yet")