# openapi.yaml
openapi: 3.0.4
info:
  description: |
    The **Tag Management System API** provides RESTful access to all tag-related operations—creation, updates, deletion, linking, and querying of tags and their relationships.
    It also supports metadata assignment for tagged entities such as images, documents, or video stills.

    ## Principles
    - **RESTful architecture** — resource-based URIs and standard HTTP verbs (GET, POST, PATCH, DELETE)
    - **JSON payloads** — all requests and responses use JSON
    - **Frontend-controlled authentication** — only frontend clients (CLI, UI) interact with the API through validated tokens (except for development/testing)
  title: Tag Management System API
  version: 0.1.1
servers:
- description: Local development
  url: http://localhost:8000/api/v1
...
paths:
...
  /tags:
    get:
      parameters:
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          example: 25
          type: integer
        style: form
      - explode: true
        in: query
        name: offset
        required: false
        schema:
          example: 0
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags_get_200_response"
          description: Paginated list of tags
      summary: List all tags
      tags:
      - tags:core
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
          description: Tag created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid input
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag already exists
      summary: Create a new tag
      tags:
      - tags:core
  /tags/{id}:
    delete:
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Tag deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: Delete a tag
      tags:
      - tags:core
    get:
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
          description: Tag found
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: Retrieve tag by ID
      tags:
      - tags:core
    patch:
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
          description: Tag updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: Update tag metadata
      tags:
      - tags:core
...
  /tags/{id}/relationships:
    delete:
      description: |
        Deletes all relationships where this tag is **Tag A** (the source).
        Use with caution — this removes every outgoing relationship for this tag.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Relationships deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: Delete relationships for a tag
      tags:
      - tags:relationships
    get:
      description: |
        Retrieve all relationships where this tag acts as **Tag A** (the dominant/source).
        Use the optional `direction` query parameter to fetch relationships where this tag
        is either the **source** or the **target**.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: |
          Controls directionality of the relationships.
          - `outgoing`: tag acts as **Tag A**
          - `incoming`: tag acts as **Tag B**
        explode: true
        in: query
        name: direction
        required: false
        schema:
          default: outgoing
          enum:
          - outgoing
          - incoming
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__relationships_get_200_response"
          description: List of tag relationships
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: List relationships for a tag
      tags:
      - tags:relationships
    patch:
      description: |
        Bulk update of relationships where this tag is the **source (Tag A)**.
        Typically used to modify multiple related tags or relationship types at once.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__relationships_get_200_response"
          description: Relationships updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: Update relationships for a tag
      tags:
      - tags:relationships
    post:
      description: |
        Creates a new relationship originating from this tag as **Tag A**.
        The request body specifies the related tag (`tag_b_id`) and the `relationship_type_id`.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagRelationshipInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagRelationship"
          description: Relationship created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid input
      summary: Create a relationship for a tag
      tags:
      - tags:relationships
  /tags/{id}/relationships/{relationship_id}:
    delete:
      description: |
        Deletes a single tag relationship record where this tag acts as **Tag A**.
        This removes the association between the two tags without deleting either tag.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: relationship_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Relationship deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Relationship not found
      summary: Delete tag relationship
      tags:
      - tags:relationships
    get:
      description: |
        Retrieves a single relationship record where this tag acts as **Tag A** (the dominant or source tag).
        This allows inspection of one connection between tags, such as "Vehicle → Car".
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Unique identifier of the tag relationship record
        explode: false
        in: path
        name: relationship_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagRelationship"
          description: Relationship found
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Relationship not found
      summary: Retrieve tag relationship
      tags:
      - tags:relationships
    patch:
      description: |
        Updates the relationship type or description of a single tag-to-tag relationship.
        The participating tags (`tag_a_id`, `tag_b_id`) remain immutable after creation.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: relationship_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_tags__id__relationships__relationship_id__patch_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagRelationship"
          description: Relationship updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Relationship not found
      summary: Update tag relationship
      tags:
      - tags:relationships
...
  /tags/{id}/ratings:
    delete:
      description: |
        Removes all contextual ratings associated with this tag.
        Use cautiously — this clears all user-assigned ratings for the tag.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Contextual ratings deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: Delete contextual ratings for a tag
      tags:
      - tags:ratings
    get:
      description: |
        Retrieve all contextual ratings applied to this tag, grouped by context.
        For example, "Dog" might have a `likeness` rating of 9/10 and a `clarity` rating of 8/10.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__ratings_get_200_response"
          description: List of contextual ratings for the tag
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: List contextual ratings for a tag
      tags:
      - tags:ratings
    patch:
      description: |
        Applies or updates contextual ratings for this tag across one or more contexts.
        Used to express subjective or objective evaluations (e.g., "Likeness", "Clarity").
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextualRatingUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__ratings_get_200_response"
          description: Contextual ratings applied or updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag or context not found
      summary: Apply or update contextual ratings for a tag
      tags:
      - tags:ratings
  /tags/{id}/ratings/{rating_id}:
    delete:
      description: |
        Deletes a single contextual rating record from this tag.
        The rating definition itself (in `/ratings`) remains unaffected.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: rating_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Contextual rating deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Contextual rating not found
      summary: Delete a contextual rating for a tag
      tags:
      - tags:ratings
    get:
      description: |
        Retrieves a specific contextual rating record for this tag, identified by its unique rating assignment ID.
        This does **not** refer to the `/ratings` table entry itself, but to the *rating instance* assigned to the tag.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Unique identifier of the contextual rating record
        explode: false
        in: path
        name: rating_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagContextualRatingInput"
          description: Contextual rating found
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Contextual rating not found
      summary: Retrieve a contextual rating for a tag
      tags:
      - tags:ratings
    patch:
      description: |
        Updates the rating, context, or user assignment for a specific contextual rating record.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: rating_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagContextualRatingInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagContextualRatingInput"
          description: Contextual rating updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Contextual rating not found
      summary: Update a contextual rating for a tag
      tags:
      - tags:ratings
  /tags/{id}/relationship_ratings:
    delete:
      description: |
        Deletes all ratings assigned to relationships where this tag acts as **Tag A**.
        This removes evaluations on all related connections for the tag.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Relationship ratings deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: Delete relationship ratings for a tag
      tags:
      - tags:relationship_ratings
    get:
      description: |
        Retrieve all ratings applied to relationships where this tag acts as **Tag A**.
        For example, "Dog → Animal" may have a `relevance` rating of 0.95.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__relationship_ratings_get_200_response"
          description: List of ratings for relationships originating from this tag
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag not found
      summary: List relationship ratings for a tag
      tags:
      - tags:relationship_ratings
    patch:
      description: |
        Applies or updates ratings on relationships where this tag acts as **Tag A**.
        Each rating reflects a subjective or objective evaluation of the connection strength or quality.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipRatingUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__relationship_ratings_patch_200_response"
          description: Relationship ratings recorded
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag or related tag not found
      summary: Apply or update relationship ratings for a tag
      tags:
      - tags:relationship_ratings
  /tags/{id}/relationship_ratings/{rating_id}:
    delete:
      description: |
        Deletes a single relationship rating record where this tag acts as **Tag A**.
        This operation removes only the rating, not the underlying tag relationship itself.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: rating_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Relationship rating deleted
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Relationship rating not found
      summary: Delete relationship rating for a tag
      tags:
      - tags:relationship_ratings
    get:
      description: |
        Retrieves a single relationship rating record where this tag acts as **Tag A** (the source).
        Each record represents a rating applied to one connection between Tag A and Tag B
        under a specific context.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Unique identifier of the relationship rating record
        explode: false
        in: path
        name: rating_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagRelationshipRatingInput"
          description: Relationship rating found
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Relationship rating not found
      summary: Retrieve relationship rating for a tag
      tags:
      - tags:relationship_ratings
    patch:
      description: |
        Updates an existing relationship rating record.
        The linked tags (Tag A, Tag B) cannot be changed after creation, but you can modify
        the `context_id` or `rating_id` to reflect a new evaluation.
      parameters:
      - description: Unique identifier for the tag
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: rating_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_tags__id__relationship_ratings__rating_id__patch_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagRelationshipRatingInput"
          description: Relationship rating updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Relationship rating not found
      summary: Update relationship rating for a tag
      tags:
      - tags:relationship_ratings
...
  /entities/{id}/tags:
    get:
      description: "Retrieve all tags currently assigned to this entity, including\
        \ contextual and metadata details."
      parameters:
      - description: Unique identifier for the entity
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_entities__id__tags_get_200_response"
          description: List of assigned tags
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Entity not found
      summary: List tags assigned to an entity
      tags:
      - entities:tags
    patch:
      description: "Assign or unassign tags for this entity, including updates to\
        \ context or metadata on existing assignments."
      parameters:
      - description: Unique identifier for the entity
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityTagUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_entities__id__tags_patch_200_response"
          description: Tag assignments updated
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Entity or tags not found
      summary: Update tags assigned to an entity
      tags:
      - entities:tags
  /entities/{id}/tags/{tag_id}:
    delete:
      description: Unassign a specific tag from this entity while leaving the tag
        itself intact in the database.
      parameters:
      - description: Unique identifier for the entity
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Unique identifier for the tag assignment
        explode: false
        in: path
        name: tag_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_entities__id__tags__tag_id__delete_200_response"
          description: Tag unassigned from entity
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Entity or tag assignment not found
      summary: Remove tag assignment from entity
      tags:
      - entities:tags
    get:
      description: "Retrieve information about a specific tag attached to this entity,\
        \ including contextual data if present."
      parameters:
      - description: Unique identifier for the entity
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Unique identifier for the tag assignment
        explode: false
        in: path
        name: tag_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityTagAssignment"
          description: Tag assignment found
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Tag assignment not found
      summary: Retrieve specific tag assignment
      tags:
      - entities:tags
...
components:
...
  schemas:
...
    TagInput:
      example:
        metadata: "{}"
        name: name
        part_of_speech_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        display_name: display_name
      properties:
        name:
          title: name
          type: string
        display_name:
          title: display_name
          type: string
        metadata:
          title: metadata
          type: object
        part_of_speech_id:
          format: uuid
          title: part_of_speech_id
          type: string
      required:
      - name
      title: TagInput
      type: object
...
    _tags_get_200_response:
      allOf:
      - $ref: "#/components/schemas/PaginatedResponse"
      - properties:
          results:
            items:
              $ref: "#/components/schemas/Tag"
            type: array
        type: object
      example:
        total: 42
        results:
        - id: a123e456-78b9-4cde-8123-456789abcdef
          name: dog
          display_name: Dog
          metadata:
            source: Label Studio
        - id: a123e456-78b9-4cde-8123-456789abcdef
          name: dog
          display_name: Dog
          metadata:
            source: Label Studio
      title: _tags_get_200_response
    _tags__id__delete_200_response:
      example:
        status: deleted
      properties:
        status:
          example: deleted
          title: status
          type: string
      title: _tags__id__delete_200_response
      type: object
...